+++ public/advanced.php
@@ -633,5 +633,6 @@ function setCookies() {
   <div id="loading-text"></div>
 </div>
 <iframe name="hiddeniframe" width=0 height=0 style="display:none;"></iframe>
+<?php require "logform.php"; ?>
 </body>
 </html>