+++ public/js/diskover.js
@@ -128,8 +128,15 @@ $(document).ready(function () {
 });
 
 // cookie functions
-function setCookie(cname, cvalue) {
-	document.cookie = cname + "=" + cvalue + ";path=/";
+function setCookie(cname, cvalue, exdays) {
+    if (exdays != '') {
+        var d = new Date();
+        d.setTime(d.getTime() + (exdays*24*60*60*1000));
+        var expires = "expires="+ d.toUTCString();
+        document.cookie = cname + "=" + cvalue + ";" + expires + ";path=/";
+    } else {
+        document.cookie = cname + "=" + cvalue + ";path=/";
+    }
 }
 
 function getCookie(cname) {