+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Diskover Web Change Log
 
+# [1.5.0-rc18] - 2018-10-21
+### added
+- pop up option on dashboard to opt in or out of sending anonymous usage stats to diskover developer, can change this later on admin page
+- pop up on dashboard to remind to help support by donation on Patreon
+
 # [1.5.0-rc17] - 2018-10-18
 ### fixed
 - elasticsearch BadRequest400Exception error (No mapping found for [hardlinks] in order to sort on) when no diskover indices and cookies on selectindices page